pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["ahmadabudames <ahmadabudamesmais1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
